pyproject.toml
[project]
name = "mcp-for-db"
version = "0.9.0"
license = { text = "MIT" }
description = "Multi-service MCP server supporting databases and AI services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=42.0.6",
    "openai>=1.83.0",
    "mcp[cli]>=1.3.0",
    "mysql-connector-python>=9.2.0",
    "python-dotenv>=0.16.0",
    "pypinyin>=0.54.0",
    "aiomysql>=0.1.0",
    "jinja2>=2.11.3",
    "PyJWT>=2.8.0",
    "sqlparse>=0.5.2",
    "contextvars>=2.0",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "numpy>=2.3.0",
    "nacos-sdk-python>=0.1.16",
    "fastapi>=0.115.12",
    "aiohttp>=3.4.0"
]

# 多个服务的控制台脚本入口
[project.scripts]
mcp_multi = "mcp_for_db.server.cli.main:stdio_entry"
mcp_mysql = "mcp_for_db.server.cli.mysql_entry:main"
mcp_dify = "mcp_for_db.server.cli.dify_entry:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["mcp_for_db"]

[tool.hatch.build.targets.wheel]
packages = ["mcp_for_db"]

[project.urls]
Homepage = "https://github.com/flyboy711/mcp-for-db"
Documentation = "https://github.com/flyboy711/mcp-for-db/blob/main/README.md"
Repository = "https://github.com/flyboy711/mcp-for-db.git"